refactor(missionary): deepen Partners page module
Keep the route default export as the public interface. Split the 2800-line
view into query/context, roster, detail, stats, and dialog adapters without
changing Gift Anonymity, Instant Navigation, or partner UX.
