+++ apps/missionary/app/donors/donors-page-client.tsx
@@ -1,14 +1,12 @@
 "use client";
-"use no memo";
 
-import { DonorsPageContent, useDonorsPageView } from "./use-donors-page-view";
+import { DonorsPageContent } from "./donors-page-content";
+import { DonorsPageViewProvider } from "./use-donors-page-view";
 
-function DonorsPageView() {
-  const viewModel = useDonorsPageView();
-
-  return <DonorsPageContent {...viewModel} />;
-}
-
-export default function DonorsPage() {
-  return <DonorsPageView />;
+export default function DonorsPageClient() {
+  return (
+    <DonorsPageViewProvider>
+      <DonorsPageContent />
+    </DonorsPageViewProvider>
+  );
 }
