+++ apps/missionary/app/donors/donors-page-dates.ts
@@ -1,0 +1,7 @@
+export function currentDisplayDate(): Date {
+  return new globalThis.Date();
+}
+
+export function parseDisplayDate(value: string | number | Date): Date {
+  return new globalThis.Date(value);
+}
