+++ apps/missionary/app/donors/use-donors-page-view.tsx
@@ -1,90 +1,9 @@
 "use client";
-"use no memo";
 
 import { useMissionaryDonorRows } from "@asym/database/hooks";
 import { invalidateSupabaseTableQuery } from "@asym/database/query-keys";
 import { useAuth } from "@asym/lib/hooks";
-import { motion, AnimatePresence } from "@asym/lib/motion";
-import { AddPartnerDialog } from "@asym/missionary/components/add-partner-dialog";
-import { PageHeader } from "@asym/ui/components/page-header";
-import {
-  Avatar,
-  AvatarFallback,
-  AvatarImage,
-} from "@asym/ui/components/shadcn/avatar";
-import { Badge } from "@asym/ui/components/shadcn/badge";
-import { Button, buttonVariants } from "@asym/ui/components/shadcn/button";
-import { Card, CardContent } from "@asym/ui/components/shadcn/card";
-import {
-  type ColumnDef,
-  DataTableColumnHeader,
-  DataTableResponsive,
-} from "@asym/ui/components/shadcn/data-table";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from "@asym/ui/components/shadcn/dialog";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuCheckboxItem,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@asym/ui/components/shadcn/dropdown-menu";
-import { Input } from "@asym/ui/components/shadcn/input";
-import { ScrollArea } from "@asym/ui/components/shadcn/scroll-area";
-import { Skeleton } from "@asym/ui/components/shadcn/skeleton";
-import {
-  Tabs,
-  TabsContent,
-  TabsList,
-  TabsTrigger,
-} from "@asym/ui/components/shadcn/tabs";
-import { Textarea } from "@asym/ui/components/shadcn/textarea";
-import { cn } from "@asym/ui/lib/utils";
 import { useQueryClient } from "@tanstack/react-query";
-import { format, formatDistanceToNow, differenceInMonths } from "date-fns";
-import {
-  Search,
-  Filter,
-  Mail,
-  Phone,
-  MapPin,
-  Plus,
-  Download,
-  Heart,
-  MessageSquare,
-  Send,
-  Copy,
-  ExternalLink,
-  Pencil,
-  User,
-  ArrowLeft,
-  ArrowUpRight,
-  ArrowDownUp,
-  Calendar,
-  Briefcase,
-  Clock,
-  AlertCircle,
-  RefreshCw,
-  MoreHorizontal,
-  Tag,
-  X,
-  Check,
-  Building2,
-  Globe,
-  Users,
-  Star,
-  Home,
-  Loader2,
-  Repeat,
-} from "lucide-react";
 import * as React from "react";
 import { toast } from "sonner";
 
@@ -93,229 +12,21 @@ import {
   type DonorActivityType,
   updateDonorTags,
 } from "./donor-mutation-client";
-import { DonorTasks } from "./donor-tasks";
 import { filterAndSortDonors, type SortOption } from "./donors-list-model";
-import {
-  AVAILABLE_TAGS,
-  formatCurrency,
-  getActivityBg,
-  getActivityIcon,
-  getGiftTypeIcon,
-  getPaymentMethodIcon,
-  getRecurringStatusBadge,
-  getStatusBadge,
-  getStatusColor,
-  getTagLabel,
-  getTagStyle,
-} from "./donors-model";
 import {
   applyDonorsStatFilter,
   createDefaultDonorsFilters,
   createDonorsPageSummary,
   formatDonorAddress,
-  getGivingHistoryRows,
   hasDonorsActiveFilters,
   removeTagSelection,
   toggleTagSelection,
   type DonorsPageSummary,
   type DonorsStatFilterType,
 } from "./donors-page-model";
-import { EditDonorDialog } from "./edit-donor-dialog";
 
-import type {
-  Activity,
-  ActivityType,
-  Address,
-  Donor,
-  RecurringStatus,
-} from "./donor-types";
+import type { Address, Donor } from "./donor-types";
 import type { Profile } from "@asym/database/types";
-
-function currentDisplayDate(): Date {
-  return new globalThis.Date();
-}
-
-function parseDisplayDate(value: string | number | Date): Date {
-  return new globalThis.Date(value);
-}
-
-const fadeInUp = {
-  initial: { opacity: 0, y: 10 },
-  animate: { opacity: 1, y: 0 },
-  exit: { opacity: 0, y: -10 },
-};
-
-const staggerContainer = {
-  animate: {
-    transition: {
-      staggerChildren: 0.05,
-    },
-  },
-};
-
-const scaleIn = {
-  initial: { opacity: 0, scale: 0.95 },
-  animate: { opacity: 1, scale: 1 },
-  exit: { opacity: 0, scale: 0.95 },
-};
-
-const slideInRight = {
-  initial: { opacity: 0, x: 20 },
-  animate: { opacity: 1, x: 0 },
-  exit: { opacity: 0, x: -20 },
-};
-
-const springTransition = {
-  type: "spring" as const,
-  stiffness: 300,
-  damping: 30,
-};
-
-const smoothTransition = {
-  duration: 0.2,
-  ease: [0.25, 0.1, 0.25, 1] as [number, number, number, number],
-};
-
-function DonorListSkeleton() {
-  return (
-    <div className="p-3 space-y-2">
-      {Array.from({ length: 12 }).map((_, i) => (
-        <motion.div
-          key={i}
-          initial={{ opacity: 0 }}
-          animate={{ opacity: 1 }}
-          transition={{ delay: i * 0.03 }}
-          className="flex items-center gap-3 p-4 rounded-2xl bg-white border border-zinc-100"
-        >
-          <Skeleton className="size-11 rounded-full" />
-          <div className="flex-1 space-y-2">
-            <Skeleton className="h-4 w-3/4" />
-            <Skeleton className="h-3 w-1/2" />
-          </div>
-        </motion.div>
-      ))}
-    </div>
-  );
-}
-
-function ErrorState({
-  message,
-  onRetry,
-}: {
-  message: string;
-  onRetry: () => void;
-}) {
-  return (
-    <motion.div
-      {...fadeInUp}
-      transition={smoothTransition}
-      className="flex flex-col items-center justify-center h-64 text-center p-6"
-    >
-      <motion.div
-        initial={{ scale: 0.8 }}
-        animate={{ scale: 1 }}
-        transition={springTransition}
-        className="size-16 bg-rose-50 rounded-2xl flex items-center justify-center mb-4 border border-rose-100"
-      >
-        <AlertCircle className="size-7 text-rose-500" />
-      </motion.div>
-      <p className="text-sm font-semibold text-zinc-900 mb-1">
-        Something went wrong
-      </p>
-      <p className="text-xs text-zinc-500 mb-4">{message}</p>
-      <motion.div whileHover={{ scale: 1.02 }} whileTap={{ scale: 0.98 }}>
-        <Button
-          variant="outline"
-          size="sm"
-          onClick={onRetry}
-          className="h-9 rounded-2xl border-zinc-200 bg-white text-[10px] font-semibold uppercase tracking-widest text-zinc-500 hover:text-zinc-900"
-        >
-          <RefreshCw className="size-3.5 mr-2" />
-          Try Again
-        </Button>
-      </motion.div>
-    </motion.div>
-  );
-}
-
-const MotionCard = motion.create(Card);
-
-function StatCard({
-  label,
-  value,
-  subtext,
-  icon: Icon,
-  iconBg,
-  iconColor,
-  onClick,
-  isActive,
-  delay = 0,
-}: {
-  label: string;
-  value: string | number;
-  subtext: string;
-  icon: React.ElementType;
-  iconBg: string;
-  iconColor: string;
-  onClick?: () => void;
-  isActive?: boolean;
-  delay?: number;
-}) {
-  const content = (
-    <MotionCard
-      initial={{ opacity: 0, y: 20 }}
-      animate={{ opacity: 1, y: 0 }}
-      transition={{ ...smoothTransition, delay }}
-      whileHover={{ y: -2, boxShadow: "0 8px 30px rgba(0,0,0,0.08)" }}
-      whileTap={onClick ? { scale: 0.98 } : undefined}
-      className={cn(
-        "border-zinc-200 bg-white shadow-sm transition-[color,background-color,border-color,box-shadow,transform,opacity] rounded-xl",
-        onClick && "cursor-pointer",
-        isActive && "border-blue-400 ring-2 ring-blue-100",
-      )}
-    >
-      <CardContent className="p-4">
-        <div className="flex items-start justify-between">
-          <div className="space-y-0.5">
-            <p className="text-[10px] font-semibold uppercase tracking-wider text-zinc-400">
-              {label}
-            </p>
-            <motion.p
-              key={value}
-              initial={{ opacity: 0, y: 5 }}
-              animate={{ opacity: 1, y: 0 }}
-              className="text-xl font-semibold tracking-tight text-zinc-900"
-            >
-              {value}
-            </motion.p>
-            <span className="text-[10px] font-medium text-zinc-400 uppercase tracking-wider">
-              {subtext}
-            </span>
-          </div>
-          <motion.div
-            whileHover={{ scale: 1.1, rotate: 5 }}
-            transition={springTransition}
-            className={cn(
-              "size-9 rounded-lg border flex items-center justify-center",
-              iconBg,
-            )}
-          >
-            <Icon className={cn("size-4", iconColor)} />
-          </motion.div>
-        </div>
-      </CardContent>
-    </MotionCard>
-  );
-
-  if (onClick) {
-    return (
-      <button onClick={onClick} className="text-left w-full">
-        {content}
-      </button>
-    );
-  }
-  return content;
-}
 
 type DonorsPageViewModel = {
   profile: Profile | null;
@@ -357,11 +68,6 @@ type DonorsPageViewModel = {
     setSortBy: (value: SortOption) => void;
     toggleSortAsc: () => void;
   };
-  tables: {
-    donorColumns: ColumnDef<Donor>[];
-    givingHistoryColumns: ColumnDef<Activity>[];
-    givingHistoryRows: Activity[];
-  };
   noteComposer: {
     isOpen: boolean;
     isSaving: boolean;
@@ -395,9 +101,12 @@ type DonorsPageViewModel = {
   };
 };
 
-export function useDonorsPageView(): DonorsPageViewModel {
+function useDonorsPageView(): DonorsPageViewModel {
   const { profile, loading: authLoading } = useAuth();
   const queryClient = useQueryClient();
+  // Gift Anonymity: missionary partner rows are server-redacted. Keep the
+  // missionary donor query; do not switch this surface to a client-side live
+  // collection of unredacted donor rows.
   const donorsQuery = useMissionaryDonorRows(profile?.id);
   const [selectedDonorId, setSelectedDonorId] = React.useState<string | null>(
     null,
@@ -469,105 +178,6 @@ export function useDonorsPageView(): DonorsPageViewModel {
   const selectedDonor = React.useMemo(
     () => donors.find((d) => d.id === selectedDonorId) || null,
     [donors, selectedDonorId],
-  );
-
-  const donorColumns = React.useMemo<ColumnDef<Donor>[]>(
-    () => [
-      {
-        accessorKey: "name",
-        header: ({ column }) => (
-          <DataTableColumnHeader column={column} title="Partner" />
-        ),
-        cell: ({ row }) => {
-          const donor = row.original;
-          const isSelected = selectedDonorId === donor.id;
-
-          return (
-            <div className="flex items-center gap-3 py-1">
-              <div className="relative shrink-0">
-                <Avatar
-                  className={cn(
-                    "size-10 border-2",
-                    isSelected ? "border-zinc-700" : "border-white shadow-sm",
-                  )}
-                >
-                  <AvatarImage src={donor.avatar_url} />
-                  <AvatarFallback
-                    className={cn(
-                      "text-xs font-semibold",
-                      isSelected
-                        ? "bg-zinc-800 text-zinc-300"
-                        : "bg-zinc-100 text-zinc-500",
-                    )}
-                  >
-                    {donor.initials}
-                  </AvatarFallback>
-                </Avatar>
-                <div
-                  className={cn(
-                    "absolute -bottom-0.5 -right-0.5 size-3 rounded-full border-2",
-                    isSelected ? "border-zinc-900" : "border-white",
-                    getStatusColor(donor.status),
-                  )}
-                />
-              </div>
-              <div className="flex-1 min-w-0">
-                <div className="flex items-center justify-between mb-0.5">
-                  <span className="font-semibold text-sm truncate text-zinc-900">
-                    {donor.name}
-                  </span>
-                  {donor.has_active_pledge && (
-                    <div
-                      className="size-2 rounded-full shrink-0 ml-1 bg-emerald-500"
-                      title="Active recurring donation"
-                    />
-                  )}
-                </div>
-                <div className="flex items-center justify-between">
-                  <span className="text-[10px] truncate max-w-[100px] font-medium uppercase tracking-wider text-zinc-400">
-                    {donor.location || "Unknown"}
-                  </span>
-                  <span className="text-xs font-semibold text-zinc-900">
-                    {formatCurrency(donor.total_given)}
-                  </span>
-                </div>
-              </div>
-            </div>
-          );
-        },
-      },
-      {
-        accessorKey: "status",
-        header: ({ column }) => (
-          <DataTableColumnHeader column={column} title="Status" />
-        ),
-        cell: ({ row }) => getStatusBadge(row.original.status),
-      },
-      {
-        accessorKey: "total_given",
-        header: ({ column }) => (
-          <div className="text-right">
-            <DataTableColumnHeader
-              className="justify-end"
-              column={column}
-              title="Given"
-            />
-          </div>
-        ),
-        cell: ({ row }) => (
-          <div className="text-right font-semibold text-zinc-900 tabular-nums">
-            {formatCurrency(row.original.total_given)}
-          </div>
-        ),
-      },
-      {
-        accessorKey: "frequency",
-        header: ({ column }) => (
-          <DataTableColumnHeader column={column} title="Frequency" />
-        ),
-      },
-    ],
-    [selectedDonorId],
   );
 
   React.useEffect(() => {
@@ -719,78 +329,6 @@ export function useDonorsPageView(): DonorsPageViewModel {
     pledgeFilter,
   });
 
-  const givingHistoryRows = React.useMemo(
-    () => getGivingHistoryRows(selectedDonor),
-    [selectedDonor],
-  );
-
-  const givingHistoryColumns = React.useMemo<ColumnDef<Activity>[]>(
-    () => [
-      {
-        accessorKey: "date",
-        header: ({ column }) => (
-          <DataTableColumnHeader column={column} title="Date" />
-        ),
-        cell: ({ row }) =>
-          format(parseDisplayDate(row.original.date), "MMM d, yyyy"),
-      },
-      {
-        accessorKey: "title",
-        header: ({ column }) => (
-          <DataTableColumnHeader column={column} title="Type" />
-        ),
-      },
-      {
-        accessorKey: "gift_type",
-        header: ({ column }) => (
-          <DataTableColumnHeader column={column} title="Method" />
-        ),
-        cell: ({ row }) => (
-          <span className="flex items-center gap-1.5 text-zinc-500">
-            {row.original.gift_type && getGiftTypeIcon(row.original.gift_type)}
-            {row.original.gift_type || "Online"}
-          </span>
-        ),
-      },
-      {
-        accessorKey: "amount",
-        header: ({ column }) => (
-          <div className="text-right">
-            <DataTableColumnHeader
-              className="justify-end"
-              column={column}
-              title="Amount"
-            />
-          </div>
-        ),
-        cell: ({ row }) => (
-          <div className="text-right font-semibold text-zinc-900">
-            {formatCurrency(row.original.amount || 0)}
-          </div>
-        ),
-      },
-      {
-        accessorKey: "status",
-        header: ({ column }) => (
-          <DataTableColumnHeader column={column} title="Status" />
-        ),
-        cell: ({ row }) => (
-          <Badge
-            className={cn(
-              "font-semibold rounded-full text-[9px] uppercase tracking-widest border-0",
-              row.original.status === "Failed"
-                ? "bg-rose-50 text-rose-600"
-                : "bg-emerald-50 text-emerald-700",
-            )}
-          >
-            {row.original.status || "Succeeded"}
-          </Badge>
-        ),
-      },
-    ],
-    [],
-  );
-
   return {
     profile,
     status: {
@@ -831,11 +369,6 @@ export function useDonorsPageView(): DonorsPageViewModel {
       setSortBy,
       toggleSortAsc,
     },
-    tables: {
-      donorColumns,
-      givingHistoryColumns,
-      givingHistoryRows,
-    },
     noteComposer: {
       isOpen: isNoteDialogOpen,
       isSaving: isSavingNote,
@@ -870,1959 +403,29 @@ export function useDonorsPageView(): DonorsPageViewModel {
   };
 }
 
-export function DonorsPageContent({
-  profile,
-  status,
-  summary,
-  tabs,
-  donors,
-  filters,
-  sorting,
-  tables,
-  noteComposer,
-  tagEditor,
-  editDialog,
-  actions,
-}: DonorsPageViewModel) {
-  const { isLoading, error } = status;
-  const {
-    activeCount,
-    activePledgeCount,
-    atRiskCount,
-    lapsedCount,
-    monthlyPledgeTotal,
-    totalGiven,
-  } = summary;
-  const { activeTab, setActiveTab } = tabs;
-  const {
-    all: donorRows,
-    filtered: filteredDonors,
-    selected: selectedDonor,
-    selectById,
-    clearSelection,
-    hasMore: hasMoreDonors,
-    isLoadingMore: isLoadingMoreDonors,
-    loadMore: loadMoreDonors,
-  } = donors;
-  const {
-    searchTerm,
-    statusFilter,
-    tagFilter,
-    pledgeFilter,
-    hasActiveFilters,
-    setSearchTerm,
-    setStatusFilter,
-    setPledgeFilter,
-    toggleTag: toggleFilterTag,
-    removeTag: removeFilterTag,
-    clearAll: clearAllFilters,
-  } = filters;
-  const { sortBy, sortAsc, setSortBy, toggleSortAsc } = sorting;
-  const { donorColumns, givingHistoryColumns, givingHistoryRows } = tables;
-  const { refreshDonors, copyToClipboard, applyStatFilter, formatAddress } =
-    actions;
+const DonorsPageViewContext = React.createContext<DonorsPageViewModel | null>(
+  null,
+);
 
+export function DonorsPageViewProvider({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  const view = useDonorsPageView();
   return (
-    <motion.div
-      initial={{ opacity: 0 }}
-      animate={{ opacity: 1 }}
-      transition={{ duration: 0.3 }}
-      className="space-y-6"
-    >
-      <PageHeader
-        title="Partners"
-        description="Manage your support network and donor relationships."
-      >
-        <motion.div whileHover={{ scale: 1.02 }} whileTap={{ scale: 0.98 }}>
-          <Button
-            variant="outline"
-            size="sm"
-            className="h-9 px-4 text-xs font-medium"
-          >
-            <Download className="mr-2 size-4" />
-            Export
-          </Button>
-        </motion.div>
-        {profile?.id && (
-          <AddPartnerDialog
-            missionaryId={profile.id}
-            onSuccess={refreshDonors}
-            trigger={
-              <Button
-                size="sm"
-                className="h-9 px-4 text-xs font-medium hover-scale-subtle"
-              >
-                <Plus className="mr-2 size-4" />
-                Add Partner
-              </Button>
-            }
-          />
-        )}
-      </PageHeader>
-
-      <div className="grid grid-cols-1 gap-4 sm:grid-cols-2 lg:grid-cols-4">
-        <StatCard
-          label={hasMoreDonors ? "Partners loaded" : "Total Partners"}
-          value={hasMoreDonors ? `${donorRows.length}+` : donorRows.length}
-          subtext={
-            hasMoreDonors
-              ? `${activeCount} active in loaded window`
-              : `${activeCount} active`
-          }
-          icon={Users}
-          iconBg="bg-zinc-50 border-zinc-100"
-          iconColor="text-zinc-900"
-          delay={0}
-        />
-        <StatCard
-          label="Total Given"
-          value={formatCurrency(totalGiven)}
-          subtext={hasMoreDonors ? "Lifetime (loaded window)" : "Lifetime"}
-          icon={Heart}
-          iconBg="bg-emerald-50 border-emerald-100"
-          iconColor="text-emerald-600"
-          delay={0.05}
-        />
-        <StatCard
-          label="Recurring Donations"
-          value={activePledgeCount}
-          subtext={
-            hasMoreDonors
-              ? `${formatCurrency(monthlyPledgeTotal)}/mo (loaded window)`
-              : `${formatCurrency(monthlyPledgeTotal)}/mo`
-          }
-          icon={Repeat}
-          iconBg="bg-blue-50 border-blue-100"
-          iconColor="text-blue-600"
-          onClick={() => applyStatFilter("activePledge")}
-          isActive={pledgeFilter === "Active"}
-          delay={0.1}
-        />
-        <StatCard
-          label="Needs Attention"
-          value={atRiskCount + lapsedCount}
-          subtext={
-            hasMoreDonors
-              ? `${atRiskCount} at risk, ${lapsedCount} lapsed (loaded window)`
-              : `${atRiskCount} at risk, ${lapsedCount} lapsed`
-          }
-          icon={AlertCircle}
-          iconBg="bg-amber-50 border-amber-100"
-          iconColor="text-amber-600"
-          onClick={() => applyStatFilter("atRisk")}
-          isActive={statusFilter === "At Risk"}
-          delay={0.15}
-        />
-      </div>
-
-      <div className="grid grid-cols-1 lg:grid-cols-12 gap-6">
-        <motion.div
-          initial={{ opacity: 0, x: -20 }}
-          animate={{ opacity: 1, x: 0 }}
-          transition={{ ...smoothTransition, delay: 0.2 }}
-          className="lg:col-span-4 xl:col-span-3"
-        >
-          <Card className="border-zinc-200 bg-white rounded-2xl overflow-hidden shadow-sm h-full flex flex-col">
-            <div className="p-4 border-b border-zinc-100 space-y-4 shrink-0">
-              <div className="flex items-center justify-between">
-                <h2 className="text-[10px] font-semibold uppercase tracking-widest text-zinc-400">
-                  Partner List{" "}
-                  {hasActiveFilters && (
-                    <span className="text-blue-600">
-                      ({filteredDonors.length})
-                    </span>
-                  )}
-                </h2>
-                <div className="flex gap-1">
-                  <DropdownMenu>
-                    <DropdownMenuTrigger
-                      render={
-                        <Button
-                          variant="ghost"
-                          size="icon"
-                          className="size-8 text-zinc-400 hover:text-zinc-900 rounded-lg"
-                        >
-                          <ArrowDownUp className="size-4" />
-                        </Button>
-                      }
-                    />
-                    <DropdownMenuContent
-                      align="end"
-                      className="w-48 rounded-xl border-zinc-100 shadow-xl"
-                    >
-                      <DropdownMenuLabel className="text-[10px] font-semibold uppercase tracking-widest text-zinc-400">
-                        Sort By
-                      </DropdownMenuLabel>
-                      <DropdownMenuSeparator className="bg-zinc-100" />
-                      {[
-                        { value: "last_gift", label: "Last Gift Date" },
-                        { value: "total_given", label: "Total Given" },
-                        { value: "name", label: "Name" },
-                        { value: "joined_date", label: "Partner Since" },
-                      ].map((opt) => (
-                        <DropdownMenuCheckboxItem
-                          key={opt.value}
-                          checked={sortBy === opt.value}
-                          onCheckedChange={() =>
-                            setSortBy(opt.value as SortOption)
-                          }
-                          className="text-xs font-medium"
-                        >
-                          {opt.label}
-                        </DropdownMenuCheckboxItem>
-                      ))}
-                      <DropdownMenuSeparator className="bg-zinc-100" />
-                      <DropdownMenuCheckboxItem
-                        checked={sortAsc}
-                        onCheckedChange={toggleSortAsc}
-                        className="text-xs font-medium"
-                      >
-                        Ascending
-                      </DropdownMenuCheckboxItem>
-                    </DropdownMenuContent>
-                  </DropdownMenu>
-                  <DropdownMenu>
-                    <DropdownMenuTrigger
-                      render={
-                        <Button
-                          variant="ghost"
-                          size="icon"
-                          className={cn(
-                            "size-8 rounded-lg",
-                            hasActiveFilters
-                              ? "text-blue-600 bg-blue-50"
-                              : "text-zinc-400 hover:text-zinc-900",
-                          )}
-                        >
-                          <Filter className="size-4" />
-                        </Button>
-                      }
-                    />
-                    <DropdownMenuContent
-                      align="end"
-                      className="w-56 rounded-xl border-zinc-100 shadow-xl max-h-[400px] overflow-y-auto"
-                    >
-                      <DropdownMenuLabel className="text-[10px] font-semibold uppercase tracking-widest text-zinc-400">
-                        Filter by Status
-                      </DropdownMenuLabel>
-                      <DropdownMenuSeparator className="bg-zinc-100" />
-                      {["All", "Active", "New", "Lapsed", "At Risk"].map(
-                        (s) => (
-                          <DropdownMenuCheckboxItem
-                            key={s}
-                            checked={statusFilter === s}
-                            onCheckedChange={() => setStatusFilter(s)}
-                            className="text-xs font-medium"
-                          >
-                            {s}
-                          </DropdownMenuCheckboxItem>
-                        ),
-                      )}
-                      <DropdownMenuSeparator className="bg-zinc-100" />
-                      <DropdownMenuLabel className="text-[10px] font-semibold uppercase tracking-widest text-zinc-400">
-                        Filter by Recurring
-                      </DropdownMenuLabel>
-                      <DropdownMenuSeparator className="bg-zinc-100" />
-                      {["All", "Active", "Inactive"].map((p) => (
-                        <DropdownMenuCheckboxItem
-                          key={p}
-                          checked={pledgeFilter === p}
-                          onCheckedChange={() => setPledgeFilter(p)}
-                          className="text-xs font-medium"
-                        >
-                          {p === "Active"
-                            ? "Has Recurring"
-                            : p === "Inactive"
-                              ? "No Recurring"
-                              : "All"}
-                        </DropdownMenuCheckboxItem>
-                      ))}
-                      <DropdownMenuSeparator className="bg-zinc-100" />
-                      <DropdownMenuLabel className="text-[10px] font-semibold uppercase tracking-widest text-zinc-400">
-                        Filter by Tag
-                      </DropdownMenuLabel>
-                      <DropdownMenuSeparator className="bg-zinc-100" />
-                      {AVAILABLE_TAGS.map((tag) => (
-                        <DropdownMenuCheckboxItem
-                          key={tag.id}
-                          checked={tagFilter.includes(tag.id)}
-                          onCheckedChange={() => toggleFilterTag(tag.id)}
-                          className="text-xs font-medium"
-                        >
-                          {tag.label}
-                        </DropdownMenuCheckboxItem>
-                      ))}
-                      {hasActiveFilters && (
-                        <>
-                          <DropdownMenuSeparator className="bg-zinc-100" />
-                          <DropdownMenuItem
-                            onClick={clearAllFilters}
-                            className="text-xs font-medium text-rose-600"
-                          >
-                            Clear All Filters
-                          </DropdownMenuItem>
-                        </>
-                      )}
-                    </DropdownMenuContent>
-                  </DropdownMenu>
-                </div>
-              </div>
-              <div className="relative">
-                <Search className="absolute left-3 top-2.5 size-4 text-zinc-400" />
-                <Input
-                  placeholder="Search partners..."
-                  className="pl-9 bg-zinc-50 border-zinc-100 focus:bg-white focus:border-zinc-300 transition-[color,background-color,border-color,box-shadow,transform,opacity] h-10 rounded-xl text-sm"
-                  value={searchTerm}
-                  onChange={(e) => setSearchTerm(e.target.value)}
-                />
-              </div>
-              <AnimatePresence mode="popLayout">
-                {hasActiveFilters && (
-                  <motion.div
-                    initial={{ opacity: 0, height: 0 }}
-                    animate={{ opacity: 1, height: "auto" }}
-                    exit={{ opacity: 0, height: 0 }}
-                    className="flex flex-wrap gap-1.5 overflow-hidden"
-                  >
-                    {statusFilter !== "All" && (
-                      <motion.div
-                        layout
-                        {...scaleIn}
-                        transition={springTransition}
-                      >
-                        <Badge
-                          variant="outline"
-                          className="text-[9px] font-semibold uppercase tracking-widest px-2 py-0.5 rounded-full bg-zinc-100 text-zinc-600 border-zinc-200"
-                        >
-                          {statusFilter}
-                          <button
-                            onClick={() => setStatusFilter("All")}
-                            className="ml-1 hover:text-zinc-900"
-                          >
-                            <X className="size-2.5" />
-                          </button>
-                        </Badge>
-                      </motion.div>
-                    )}
-                    {pledgeFilter !== "All" && (
-                      <motion.div
-                        layout
-                        {...scaleIn}
-                        transition={springTransition}
-                      >
-                        <Badge
-                          variant="outline"
-                          className="text-[9px] font-semibold uppercase tracking-widest px-2 py-0.5 rounded-full bg-blue-50 text-blue-600 border-blue-200"
-                        >
-                          {pledgeFilter === "Active"
-                            ? "Recurring"
-                            : "No Recurring"}
-                          <button
-                            onClick={() => setPledgeFilter("All")}
-                            className="ml-1 hover:text-blue-900"
-                          >
-                            <X className="size-2.5" />
-                          </button>
-                        </Badge>
-                      </motion.div>
-                    )}
-                    {tagFilter.map((tag) => (
-                      <motion.div
-                        key={tag}
-                        layout
-                        {...scaleIn}
-                        transition={springTransition}
-                      >
-                        <Badge
-                          variant="outline"
-                          className={cn(
-                            "text-[9px] font-semibold uppercase tracking-widest px-2 py-0.5 rounded-full border",
-                            getTagStyle(tag),
-                          )}
-                        >
-                          {getTagLabel(tag)}
-                          <button
-                            onClick={() => removeFilterTag(tag)}
-                            className="ml-1"
-                          >
-                            <X className="size-2.5" />
-                          </button>
-                        </Badge>
-                      </motion.div>
-                    ))}
-                    <motion.button
-                      layout
-                      whileHover={{ scale: 1.02 }}
-                      whileTap={{ scale: 0.98 }}
-                      onClick={clearAllFilters}
-                      className="text-[9px] font-semibold uppercase tracking-widest text-rose-500 hover:text-rose-700 px-2"
-                    >
-                      Clear All
-                    </motion.button>
-                  </motion.div>
-                )}
-              </AnimatePresence>
-            </div>
-
-            <div className="flex-1 min-h-0">
-              <ScrollArea className="h-full">
-                {error ? (
-                  <ErrorState message={error} onRetry={refreshDonors} />
-                ) : isLoading ? (
-                  <DonorListSkeleton />
-                ) : filteredDonors.length === 0 ? (
-                  <motion.div
-                    {...fadeInUp}
-                    transition={smoothTransition}
-                    className="flex flex-col items-center justify-center h-64 text-center p-6"
-                  >
-                    <motion.div
-                      initial={{ scale: 0.8 }}
-                      animate={{ scale: 1 }}
-                      transition={springTransition}
-                      className="size-14 bg-zinc-100 rounded-2xl flex items-center justify-center mb-4"
-                    >
-                      <Search className="size-6 text-zinc-300" />
-                    </motion.div>
-                    <p className="text-sm font-semibold text-zinc-900">
-                      No partners found
-                    </p>
-                    <p className="text-xs text-zinc-400 mt-1">
-                      {hasActiveFilters
-                        ? "Try adjusting your filters"
-                        : "Add your first partner to get started"}
-                    </p>
-                    {hasActiveFilters && (
-                      <motion.div
-                        whileHover={{ scale: 1.02 }}
-                        whileTap={{ scale: 0.98 }}
-                      >
-                        <Button
-                          variant="outline"
-                          size="sm"
-                          onClick={clearAllFilters}
-                          className="mt-4 h-8 rounded-xl text-xs"
-                        >
-                          Clear Filters
-                        </Button>
-                      </motion.div>
-                    )}
-                  </motion.div>
-                ) : (
-                  <DataTableResponsive
-                    columns={donorColumns}
-                    data={filteredDonors}
-                    config={{
-                      enableRowSelection: false,
-                      enableColumnVisibility: true,
-                      enablePagination: true,
-                      enableFilters: false,
-                      enableSorting: false,
-                      virtualization: {
-                        enabled: true,
-                        estimateSize: 88,
-                        overscan: 10,
-                        containerHeight: 640,
-                      },
-                    }}
-                    mobileCardConfig={{
-                      primaryField: "name",
-                      secondaryField: "location",
-                      badgeField: "status",
-                    }}
-                    onRowClick={(row) => selectById(row.original.id)}
-                    emptyState={
-                      <div className="flex flex-col items-center justify-center py-12 text-center">
-                        <p className="text-sm font-semibold text-zinc-900">
-                          No partners found
-                        </p>
-                        <p className="text-xs text-zinc-400 mt-1">
-                          {hasActiveFilters
-                            ? "Try adjusting your filters."
-                            : "Add your first partner to get started."}
-                        </p>
-                      </div>
-                    }
-                  />
-                )}
-              </ScrollArea>
-            </div>
-            {hasMoreDonors && !error && !isLoading && (
-              <div className="border-t border-zinc-100 p-3 shrink-0">
-                <Button
-                  variant="outline"
-                  size="sm"
-                  onClick={loadMoreDonors}
-                  disabled={isLoadingMoreDonors}
-                  className="w-full h-9 rounded-xl text-[10px] font-semibold uppercase tracking-widest text-zinc-500 hover:text-zinc-900"
-                >
-                  {isLoadingMoreDonors ? (
-                    <>
-                      <Loader2 className="size-3.5 mr-2 animate-spin" />
-                      Loading partners
-                    </>
-                  ) : (
-                    "Load more partners"
-                  )}
-                </Button>
-              </div>
-            )}
-          </Card>
-        </motion.div>
-
-        <motion.div
-          initial={{ opacity: 0, x: 20 }}
-          animate={{ opacity: 1, x: 0 }}
-          transition={{ ...smoothTransition, delay: 0.25 }}
-          className="lg:col-span-8 xl:col-span-9"
-        >
-          <AnimatePresence mode="wait">
-            {selectedDonor ? (
-              <motion.div
-                key={selectedDonor.id}
-                {...slideInRight}
-                transition={smoothTransition}
-              >
-                <Card className="border-zinc-200 bg-white rounded-2xl overflow-hidden shadow-sm h-full flex flex-col">
-                  <div className="border-b border-zinc-100 p-6 shrink-0">
-                    <div className="flex flex-col sm:flex-row items-start sm:items-center justify-between gap-4">
-                      <div className="flex items-center gap-4">
-                        <motion.div
-                          whileHover={{ scale: 1.02 }}
-                          whileTap={{ scale: 0.98 }}
-                        >
-                          <Button
-                            variant="ghost"
-                            size="icon"
-                            className="lg:hidden size-9 text-zinc-400 rounded-xl hover:bg-zinc-100"
-                            onClick={clearSelection}
-                          >
-                            <ArrowLeft className="size-5" />
-                          </Button>
-                        </motion.div>
-                        <motion.div
-                          initial={{ scale: 0.8, opacity: 0 }}
-                          animate={{ scale: 1, opacity: 1 }}
-                          transition={springTransition}
-                        >
-                          <Avatar className="size-14 border-2 border-white shadow-lg rounded-2xl">
-                            <AvatarImage src={selectedDonor.avatar_url} />
-                            <AvatarFallback className="text-lg font-semibold bg-zinc-100 text-zinc-500">
-                              {selectedDonor.initials}
-                            </AvatarFallback>
-                          </Avatar>
-                        </motion.div>
-                        <motion.div
-                          initial={{ opacity: 0, x: -10 }}
-                          animate={{ opacity: 1, x: 0 }}
-                          transition={{ ...smoothTransition, delay: 0.1 }}
-                        >
-                          <div className="flex items-center gap-3 mb-1">
-                            <h2 className="text-lg font-semibold text-zinc-900 tracking-tight">
-                              {selectedDonor.name}
-                            </h2>
-                            {getStatusBadge(selectedDonor.status)}
-                          </div>
-                          <div className="flex items-center gap-3 text-xs text-zinc-500">
-                            <span className="flex items-center gap-1">
-                              <MapPin className="size-3" />{" "}
-                              {selectedDonor.location || "Unknown"}
-                            </span>
-                            <span className="flex items-center gap-1 capitalize">
-                              {selectedDonor.type === "Church" ? (
-                                <Building2 className="size-3" />
-                              ) : selectedDonor.type === "Organization" ? (
-                                <Briefcase className="size-3" />
-                              ) : (
-                                <User className="size-3" />
-                              )}
-                              {selectedDonor.type}
-                            </span>
-                          </div>
-                        </motion.div>
-                      </div>
-                      <motion.div
-                        initial={{ opacity: 0, y: 10 }}
-                        animate={{ opacity: 1, y: 0 }}
-                        transition={{ ...smoothTransition, delay: 0.15 }}
-                        className="flex items-center gap-2 w-full sm:w-auto"
-                      >
-                        <motion.div
-                          whileHover={{ scale: 1.02 }}
-                          whileTap={{ scale: 0.98 }}
-                          className="flex-1 sm:flex-none"
-                        >
-                          <Button
-                            variant="outline"
-                            size="sm"
-                            className="w-full h-9 px-4 text-xs font-medium rounded-xl border-zinc-200 hover:bg-zinc-50"
-                            onClick={() => noteComposer.open("note")}
-                          >
-                            <Pencil className="size-3.5 mr-1.5" /> Note
-                          </Button>
-                        </motion.div>
-                        <motion.div
-                          whileHover={{ scale: 1.02 }}
-                          whileTap={{ scale: 0.98 }}
-                          className="flex-1 sm:flex-none"
-                        >
-                          <a
-                            href={`tel:${selectedDonor.phone || selectedDonor.mobile}`}
-                            className={cn(
-                              buttonVariants({
-                                variant: "outline",
-                                size: "sm",
-                              }),
-                              "w-full h-9 px-4 text-xs font-medium rounded-xl border-zinc-200 hover:bg-zinc-50",
-                            )}
-                          >
-                            <Phone className="size-3.5 mr-1.5" /> Call
-                          </a>
-                        </motion.div>
-                        <motion.div
-                          whileHover={{ scale: 1.02 }}
-                          whileTap={{ scale: 0.98 }}
-                          className="flex-1 sm:flex-none"
-                        >
-                          <a
-                            href={`mailto:${selectedDonor.email}`}
-                            className={cn(
-                              buttonVariants({ size: "sm" }),
-                              "w-full h-9 px-4 text-xs font-medium rounded-xl",
-                            )}
-                          >
-                            <Mail className="size-3.5 mr-1.5" /> Email
-                          </a>
-                        </motion.div>
-                        <DropdownMenu>
-                          <DropdownMenuTrigger
-                            render={
-                              <Button
-                                variant="ghost"
-                                size="icon"
-                                className="size-9 text-zinc-400 rounded-xl hover:bg-zinc-100"
-                              >
-                                <MoreHorizontal className="size-5" />
-                              </Button>
-                            }
-                          />
-                          <DropdownMenuContent
-                            align="end"
-                            className="rounded-xl border-zinc-100 shadow-xl"
-                          >
-                            <DropdownMenuLabel className="text-[10px] font-semibold uppercase tracking-widest text-zinc-400">
-                              Actions
-                            </DropdownMenuLabel>
-                            <DropdownMenuSeparator className="bg-zinc-100" />
-                            <DropdownMenuItem
-                              onClick={editDialog.open}
-                              className="text-xs font-medium"
-                            >
-                              <Pencil className="size-3.5 mr-2" /> Edit Profile
-                            </DropdownMenuItem>
-                            <DropdownMenuItem
-                              onClick={tagEditor.open}
-                              className="text-xs font-medium"
-                            >
-                              <Tag className="size-3.5 mr-2" /> Manage Tags
-                            </DropdownMenuItem>
-                            <DropdownMenuSeparator className="bg-zinc-100" />
-                            <DropdownMenuItem
-                              onClick={() => noteComposer.open("call")}
-                              className="text-xs font-medium"
-                            >
-                              <Phone className="size-3.5 mr-2" /> Log Call
-                            </DropdownMenuItem>
-                            <DropdownMenuItem
-                              onClick={() => noteComposer.open("meeting")}
-                              className="text-xs font-medium"
-                            >
-                              <Briefcase className="size-3.5 mr-2" /> Log
-                              Meeting
-                            </DropdownMenuItem>
-                            <DropdownMenuItem
-                              onClick={() => noteComposer.open("email")}
-                              className="text-xs font-medium"
-                            >
-                              <Mail className="size-3.5 mr-2" /> Log Email
-                            </DropdownMenuItem>
-                          </DropdownMenuContent>
-                        </DropdownMenu>
-                      </motion.div>
-                    </div>
-
-                    <motion.div
-                      variants={staggerContainer}
-                      initial="initial"
-                      animate="animate"
-                      className="grid grid-cols-2 md:grid-cols-4 gap-3 mt-6"
-                    >
-                      {[
-                        {
-                          label: "Lifetime",
-                          value: formatCurrency(selectedDonor.total_given),
-                        },
-                        {
-                          label: "Last Gift",
-                          value: formatCurrency(selectedDonor.last_gift_amount),
-                          extra: selectedDonor.last_gift_date
-                            ? formatDistanceToNow(
-                                parseDisplayDate(selectedDonor.last_gift_date),
-                                { addSuffix: true },
-                              )
-                            : null,
-                          showPulse:
-                            selectedDonor.last_gift_date &&
-                            differenceInMonths(
-                              currentDisplayDate(),
-                              parseDisplayDate(selectedDonor.last_gift_date),
-                            ) < 1,
-                        },
-                        {
-                          label: "Frequency",
-                          value: selectedDonor.frequency || "N/A",
-                          icon: ArrowUpRight,
-                        },
-                        {
-                          label: "Partner Since",
-                          value: selectedDonor.joined_date
-                            ? format(
-                                parseDisplayDate(selectedDonor.joined_date),
-                                "MMM yyyy",
-                              )
-                            : "N/A",
-                        },
-                      ].map((stat, i) => (
-                        <motion.div
-                          key={stat.label}
-                          variants={fadeInUp}
-                          transition={{
-                            ...smoothTransition,
-                            delay: 0.2 + i * 0.05,
-                          }}
-                          whileHover={{ y: -2 }}
-                          className="bg-zinc-50 p-4 rounded-2xl border border-zinc-100"
-                        >
-                          <p className="text-[10px] font-semibold uppercase tracking-wider text-zinc-400 mb-1">
-                            {stat.label}
-                          </p>
-                          <div className="flex items-center gap-2">
-                            {stat.icon && (
-                              <stat.icon className="size-3.5 text-emerald-600" />
-                            )}
-                            <p
-                              className={cn(
-                                stat.label === "Lifetime" ||
-                                  stat.label === "Last Gift"
-                                  ? "text-lg"
-                                  : "text-sm",
-                                "font-semibold text-zinc-900",
-                              )}
-                            >
-                              {stat.value}
-                            </p>
-                            {stat.showPulse && (
-                              <motion.div
-                                animate={{
-                                  scale: [1, 1.3, 1],
-                                  opacity: [1, 0.7, 1],
-                                }}
-                                transition={{ duration: 1.5, repeat: Infinity }}
-                                className="size-2 bg-emerald-500 rounded-full"
-                              />
-                            )}
-                          </div>
-                          {stat.extra && (
-                            <p className="text-[10px] text-zinc-400 mt-0.5">
-                              {stat.extra}
-                            </p>
-                          )}
-                        </motion.div>
-                      ))}
-                    </motion.div>
-
-                    <motion.div
-                      initial={{ opacity: 0 }}
-                      animate={{ opacity: 1 }}
-                      transition={{ delay: 0.4 }}
-                      className="flex flex-wrap items-center gap-1.5 mt-4"
-                    >
-                      <AnimatePresence mode="popLayout">
-                        {(selectedDonor.tags || []).map((tag, i) => (
-                          <motion.div
-                            key={tag}
-                            layout
-                            initial={{ opacity: 0, scale: 0.8 }}
-                            animate={{ opacity: 1, scale: 1 }}
-                            exit={{ opacity: 0, scale: 0.8 }}
-                            transition={{
-                              ...springTransition,
-                              delay: i * 0.03,
-                            }}
-                          >
-                            <Badge
-                              variant="outline"
-                              className={cn(
-                                "text-[9px] font-semibold uppercase tracking-widest px-2 py-0.5 rounded-full border",
-                                getTagStyle(tag),
-                              )}
-                            >
-                              {getTagLabel(tag)}
-                            </Badge>
-                          </motion.div>
-                        ))}
-                      </AnimatePresence>
-                      <motion.div
-                        whileHover={{ scale: 1.02 }}
-                        whileTap={{ scale: 0.98 }}
-                      >
-                        <Button
-                          variant="ghost"
-                          size="sm"
-                          className="h-6 px-2 text-[10px] font-semibold uppercase tracking-widest text-zinc-400 hover:text-zinc-900"
-                          onClick={tagEditor.open}
-                        >
-                          <Plus className="size-3 mr-1" /> Add Tag
-                        </Button>
-                      </motion.div>
-                    </motion.div>
-                  </div>
-
-                  <Tabs
-                    value={activeTab}
-                    onValueChange={setActiveTab}
-                    className="flex-1 flex flex-col min-h-0"
-                  >
-                    <div className="px-6 py-4 border-b border-zinc-100 shrink-0">
-                      <TabsList className="bg-zinc-100/50 border border-zinc-100 p-1.5 h-auto rounded-2xl w-full sm:w-auto grid grid-cols-5 sm:flex">
-                        {[
-                          "overview",
-                          "tasks",
-                          "contact",
-                          "recurring",
-                          "giving",
-                        ].map((tab) => (
-                          <TabsTrigger
-                            key={tab}
-                            value={tab}
-                            className="rounded-xl data-active:bg-white data-active:shadow-sm px-4 sm:px-6 py-2 text-[10px] font-semibold uppercase tracking-widest text-zinc-400 data-active:text-zinc-900 transition-colors"
-                          >
-                            {tab === "overview"
-                              ? "Overview"
-                              : tab === "tasks"
-                                ? "Tasks"
-                                : tab === "contact"
-                                  ? "Contact"
-                                  : tab === "recurring"
-                                    ? "Recurring"
-                                    : "Giving"}
-                          </TabsTrigger>
-                        ))}
-                      </TabsList>
-                    </div>
-
-                    <ScrollArea className="flex-1 min-h-0">
-                      <div className="p-6">
-                        <AnimatePresence mode="wait">
-                          <TabsContent
-                            value="overview"
-                            className="mt-0 space-y-6"
-                          >
-                            <motion.div
-                              {...fadeInUp}
-                              transition={smoothTransition}
-                              className="bg-zinc-50 p-4 rounded-2xl border border-zinc-100"
-                            >
-                              <Textarea
-                                placeholder="Log a call, meeting notes, or observation..."
-                                className="min-h-[80px] border-none bg-white focus:ring-0 resize-none text-sm p-3 rounded-xl shadow-sm"
-                                value={noteComposer.noteInput}
-                                onChange={(e) =>
-                                  noteComposer.setNoteInput(e.target.value)
-                                }
-                              />
-                              <div className="flex justify-between items-center mt-3 pt-3 border-t border-zinc-100">
-                                <div className="flex gap-2">
-                                  {[
-                                    {
-                                      type: "call",
-                                      icon: Phone,
-                                      bg: "bg-blue-50 text-blue-600",
-                                    },
-                                    {
-                                      type: "meeting",
-                                      icon: Briefcase,
-                                      bg: "bg-emerald-50 text-emerald-600",
-                                    },
-                                    {
-                                      type: "note",
-                                      icon: MessageSquare,
-                                      bg: "bg-zinc-200 text-zinc-700",
-                                      hidden: "hidden sm:flex",
-                                    },
-                                  ].map(({ type, icon: Icon, bg, hidden }) => (
-                                    <motion.div
-                                      key={type}
-                                      whileHover={{ scale: 1.02 }}
-                                      whileTap={{ scale: 0.98 }}
-                                    >
-                                      <Button
-                                        variant="ghost"
-                                        size="sm"
-                                        className={cn(
-                                          "h-8 rounded-lg text-[10px] font-semibold uppercase tracking-widest",
-                                          hidden,
-                                          noteComposer.activityType === type
-                                            ? bg
-                                            : "text-zinc-400 hover:text-zinc-900 hover:bg-zinc-100",
-                                        )}
-                                        onClick={() =>
-                                          noteComposer.setActivityType(
-                                            type as DonorActivityType,
-                                          )
-                                        }
-                                      >
-                                        <Icon className="size-3.5 mr-1.5" />{" "}
-                                        {type.charAt(0).toUpperCase() +
-                                          type.slice(1)}
-                                      </Button>
-                                    </motion.div>
-                                  ))}
-                                </div>
-                                <motion.div
-                                  whileHover={{ scale: 1.02 }}
-                                  whileTap={{ scale: 0.98 }}
-                                >
-                                  <Button
-                                    size="sm"
-                                    className="h-8 rounded-xl px-4 text-[10px] font-semibold uppercase tracking-widest"
-                                    onClick={noteComposer.save}
-                                    disabled={
-                                      !noteComposer.noteInput.trim() ||
-                                      noteComposer.isSaving
-                                    }
-                                  >
-                                    {noteComposer.isSaving ? (
-                                      <Loader2 className="size-3 animate-spin" />
-                                    ) : (
-                                      <>
-                                        Post <Send className="size-3 ml-1.5" />
-                                      </>
-                                    )}
-                                  </Button>
-                                </motion.div>
-                              </div>
-                            </motion.div>
-
-                            <div className="space-y-4 relative">
-                              <div className="absolute left-4 top-0 bottom-0 w-0.5 bg-zinc-100" />
-
-                              {selectedDonor.activities.length === 0 ? (
-                                <motion.div
-                                  {...fadeInUp}
-                                  className="flex flex-col items-center justify-center py-16 text-center ml-8"
-                                >
-                                  <motion.div
-                                    initial={{ scale: 0.8 }}
-                                    animate={{ scale: 1 }}
-                                    transition={springTransition}
-                                    className="size-16 bg-zinc-100 rounded-2xl flex items-center justify-center mb-4"
-                                  >
-                                    <Calendar className="size-7 text-zinc-300" />
-                                  </motion.div>
-                                  <p className="text-sm font-semibold text-zinc-900">
-                                    No activity recorded yet
-                                  </p>
-                                  <p className="text-xs text-zinc-400 mt-1">
-                                    Start by logging your first interaction
-                                  </p>
-                                </motion.div>
-                              ) : (
-                                <AnimatePresence>
-                                  {selectedDonor.activities.map(
-                                    (activity, i) => (
-                                      <motion.div
-                                        key={activity.id}
-                                        initial={{ opacity: 0, x: -20 }}
-                                        animate={{ opacity: 1, x: 0 }}
-                                        transition={{
-                                          ...smoothTransition,
-                                          delay: i * 0.05,
-                                        }}
-                                        className="relative pl-10 group"
-                                      >
-                                        <motion.div
-                                          whileHover={{ scale: 1.15 }}
-                                          className={cn(
-                                            "absolute left-0 top-1 size-8 rounded-xl flex items-center justify-center shadow-sm z-10",
-                                            getActivityBg(
-                                              activity.type as ActivityType,
-                                            ),
-                                          )}
-                                        >
-                                          {getActivityIcon(
-                                            activity.type as ActivityType,
-                                          )}
-                                        </motion.div>
-
-                                        <motion.div
-                                          whileHover={{
-                                            y: -2,
-                                            boxShadow:
-                                              "0 8px 30px rgba(0,0,0,0.08)",
-                                          }}
-                                          className="bg-white p-4 rounded-2xl border border-zinc-200 hover:border-zinc-300 transition-[color,background-color,border-color,box-shadow,transform,opacity]"
-                                        >
-                                          <div className="flex flex-col sm:flex-row sm:items-start justify-between gap-2 mb-1">
-                                            <div className="space-y-1">
-                                              <div className="flex flex-wrap items-center gap-2">
-                                                <span className="text-sm font-semibold text-zinc-900">
-                                                  {activity.title}
-                                                </span>
-                                                {activity.amount && (
-                                                  <Badge
-                                                    className={cn(
-                                                      "font-semibold px-2 h-5 rounded-lg text-[9px] uppercase tracking-widest border-0",
-                                                      activity.status ===
-                                                        "Failed"
-                                                        ? "bg-rose-50 text-rose-600"
-                                                        : "bg-emerald-50 text-emerald-700",
-                                                    )}
-                                                  >
-                                                    {formatCurrency(
-                                                      activity.amount,
-                                                    )}
-                                                  </Badge>
-                                                )}
-                                                {activity.gift_type && (
-                                                  <span className="flex items-center gap-1 text-[10px] font-medium text-zinc-400">
-                                                    {getGiftTypeIcon(
-                                                      activity.gift_type,
-                                                    )}
-                                                    {activity.gift_type}
-                                                  </span>
-                                                )}
-                                                {activity.status ===
-                                                  "Failed" && (
-                                                  <Badge className="bg-rose-50 text-rose-600 border-0 text-[9px] font-semibold uppercase tracking-widest">
-                                                    Failed
-                                                  </Badge>
-                                                )}
-                                              </div>
-                                              {activity.description && (
-                                                <p className="text-sm text-zinc-500 leading-relaxed">
-                                                  {activity.description}
-                                                </p>
-                                              )}
-                                              {activity.note && (
-                                                <p className="text-xs text-zinc-400 italic">
-                                                  {activity.note}
-                                                </p>
-                                              )}
-                                            </div>
-                                            <span className="text-[10px] font-semibold uppercase tracking-widest text-zinc-400 whitespace-nowrap">
-                                              {format(
-                                                parseDisplayDate(activity.date),
-                                                "MMM d, yyyy",
-                                              )}
-                                            </span>
-                                          </div>
-                                        </motion.div>
-                                      </motion.div>
-                                    ),
-                                  )}
-                                </AnimatePresence>
-                              )}
-                            </div>
-                          </TabsContent>
-
-                          <TabsContent value="tasks" className="mt-0 space-y-6">
-                            <DonorTasks
-                              donorId={selectedDonor.id}
-                              donorName={selectedDonor.name}
-                            />
-                          </TabsContent>
-
-                          <TabsContent
-                            value="contact"
-                            className="mt-0 space-y-6"
-                          >
-                            <motion.div
-                              {...fadeInUp}
-                              transition={smoothTransition}
-                              className="flex items-center justify-between mb-2"
-                            >
-                              <h3 className="text-sm font-semibold text-zinc-900">
-                                Contact Information
-                              </h3>
-                              <motion.div
-                                whileHover={{ scale: 1.02 }}
-                                whileTap={{ scale: 0.98 }}
-                              >
-                                <Button
-                                  variant="outline"
-                                  size="sm"
-                                  onClick={editDialog.open}
-                                  className="h-8 px-3 text-xs rounded-xl border-zinc-200"
-                                >
-                                  <Pencil className="size-3.5 mr-1.5" /> Edit
-                                </Button>
-                              </motion.div>
-                            </motion.div>
-
-                            <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
-                              <motion.div
-                                variants={staggerContainer}
-                                initial="initial"
-                                animate="animate"
-                                className="space-y-3"
-                              >
-                                {[
-                                  {
-                                    icon: Mail,
-                                    label: "Email",
-                                    value: selectedDonor.email,
-                                    preferred:
-                                      selectedDonor.preferred_contact ===
-                                      "email",
-                                    color: "blue",
-                                  },
-                                  {
-                                    icon: Phone,
-                                    label: "Primary Phone",
-                                    value: selectedDonor.phone,
-                                    preferred:
-                                      selectedDonor.preferred_contact ===
-                                      "phone",
-                                    color: "emerald",
-                                  },
-                                  {
-                                    icon: MessageSquare,
-                                    label: "Mobile / Text",
-                                    value: selectedDonor.mobile,
-                                    preferred:
-                                      selectedDonor.preferred_contact ===
-                                      "text",
-                                    color: "purple",
-                                  },
-                                  {
-                                    icon: Briefcase,
-                                    label: "Work Phone",
-                                    value: selectedDonor.work_phone,
-                                    color: "zinc",
-                                  },
-                                ].map((item, i) => (
-                                  <motion.div
-                                    key={item.label}
-                                    variants={fadeInUp}
-                                    transition={{ delay: i * 0.05 }}
-                                    whileHover={{ y: -2 }}
-                                    className="flex items-center justify-between p-4 bg-zinc-50 rounded-2xl border border-zinc-100 group hover:border-zinc-200 transition-[color,background-color,border-color,box-shadow,transform,opacity]"
-                                  >
-                                    <div className="flex items-center gap-3">
-                                      <div
-                                        className={cn(
-                                          "size-10 rounded-xl flex items-center justify-center shrink-0",
-                                          `bg-${item.color}-50 text-${item.color}-600`,
-                                        )}
-                                      >
-                                        <item.icon className="size-4" />
-                                      </div>
-                                      <div className="min-w-0">
-                                        <div className="flex items-center gap-2">
-                                          <p className="text-[10px] font-semibold uppercase tracking-widest text-zinc-400">
-                                            {item.label}
-                                          </p>
-                                          {item.preferred && (
-                                            <Badge
-                                              className={cn(
-                                                `bg-${item.color}-50 text-${item.color}-600`,
-                                                "border-0 text-[8px] font-semibold uppercase tracking-widest px-1.5 py-0",
-                                              )}
-                                            >
-                                              Preferred
-                                            </Badge>
-                                          )}
-                                        </div>
-                                        <p className="text-sm font-medium text-zinc-900 truncate">
-                                          {item.value || "Not provided"}
-                                        </p>
-                                      </div>
-                                    </div>
-                                    {item.value && (
-                                      <motion.div
-                                        whileHover={{ scale: 1.1 }}
-                                        whileTap={{ scale: 0.97 }}
-                                      >
-                                        <Button
-                                          variant="ghost"
-                                          size="icon"
-                                          className={cn(
-                                            "size-9 rounded-xl shrink-0",
-                                            `text-zinc-400 hover:text-${item.color}-600 hover:bg-${item.color}-50`,
-                                          )}
-                                          onClick={() =>
-                                            copyToClipboard(
-                                              item.value!,
-                                              item.label,
-                                            )
-                                          }
-                                        >
-                                          <Copy className="size-4" />
-                                        </Button>
-                                      </motion.div>
-                                    )}
-                                  </motion.div>
-                                ))}
-                                {selectedDonor.website && (
-                                  <motion.div
-                                    variants={fadeInUp}
-                                    whileHover={{ y: -2 }}
-                                    className="flex items-center justify-between p-4 bg-zinc-50 rounded-2xl border border-zinc-100 group hover:border-zinc-200 transition-[color,background-color,border-color,box-shadow,transform,opacity]"
-                                  >
-                                    <div className="flex items-center gap-3">
-                                      <div className="size-10 rounded-xl bg-primary/10 text-primary flex items-center justify-center shrink-0">
-                                        <Globe className="size-4" />
-                                      </div>
-                                      <div className="min-w-0">
-                                        <p className="text-[10px] font-semibold uppercase tracking-widest text-zinc-400">
-                                          Website
-                                        </p>
-                                        <p className="text-sm font-medium text-zinc-900 truncate">
-                                          {selectedDonor.website}
-                                        </p>
-                                      </div>
-                                    </div>
-                                    <motion.div
-                                      whileHover={{ scale: 1.1 }}
-                                      whileTap={{ scale: 0.97 }}
-                                    >
-                                      <a
-                                        href={
-                                          selectedDonor.website.startsWith(
-                                            "http",
-                                          )
-                                            ? selectedDonor.website
-                                            : `https://${selectedDonor.website}`
-                                        }
-                                        target="_blank"
-                                        rel="noopener noreferrer"
-                                        className={cn(
-                                          buttonVariants({
-                                            variant: "ghost",
-                                            size: "icon",
-                                          }),
-                                          "size-9 text-zinc-400 hover:text-primary hover:bg-primary/10 rounded-xl shrink-0",
-                                        )}
-                                      >
-                                        <ExternalLink className="size-4" />
-                                      </a>
-                                    </motion.div>
-                                  </motion.div>
-                                )}
-                              </motion.div>
-
-                              <motion.div
-                                variants={staggerContainer}
-                                initial="initial"
-                                animate="animate"
-                                className="space-y-4"
-                              >
-                                <motion.div
-                                  variants={fadeInUp}
-                                  whileHover={{ y: -2 }}
-                                  className="p-4 bg-zinc-50 rounded-2xl border border-zinc-100"
-                                >
-                                  <div className="flex items-start justify-between">
-                                    <div className="flex items-start gap-3">
-                                      <div className="size-10 rounded-xl bg-zinc-100 text-zinc-500 flex items-center justify-center shrink-0">
-                                        <Home className="size-4" />
-                                      </div>
-                                      <div>
-                                        <p className="text-[10px] font-semibold uppercase tracking-widest text-zinc-400 mb-1">
-                                          Mailing Address
-                                        </p>
-                                        {selectedDonor.address?.street ? (
-                                          <>
-                                            {formatAddress(
-                                              selectedDonor.address,
-                                            ).map((line, i) => (
-                                              <p
-                                                key={`${line}-${selectedDonor.id}`}
-                                                className={cn(
-                                                  "text-sm",
-                                                  i === 0
-                                                    ? "font-medium text-zinc-900"
-                                                    : "text-zinc-500",
-                                                )}
-                                              >
-                                                {line}
-                                              </p>
-                                            ))}
-                                          </>
-                                        ) : (
-                                          <p className="text-sm text-zinc-400 italic">
-                                            No address on file
-                                          </p>
-                                        )}
-                                      </div>
-                                    </div>
-                                    {selectedDonor.address?.street && (
-                                      <motion.div
-                                        whileHover={{ scale: 1.1 }}
-                                        whileTap={{ scale: 0.97 }}
-                                      >
-                                        <a
-                                          href={`https://maps.google.com/?q=${encodeURIComponent(formatAddress(selectedDonor.address).join(", "))}`}
-                                          target="_blank"
-                                          rel="noopener noreferrer"
-                                          className={cn(
-                                            buttonVariants({
-                                              variant: "ghost",
-                                              size: "icon",
-                                            }),
-                                            "size-9 text-zinc-400 hover:text-zinc-900 hover:bg-zinc-100 rounded-xl shrink-0",
-                                          )}
-                                        >
-                                          <ExternalLink className="size-4" />
-                                        </a>
-                                      </motion.div>
-                                    )}
-                                  </div>
-                                </motion.div>
-
-                                {(selectedDonor.organization ||
-                                  selectedDonor.title) && (
-                                  <motion.div
-                                    variants={fadeInUp}
-                                    whileHover={{ y: -2 }}
-                                    className="p-4 bg-zinc-50 rounded-2xl border border-zinc-100"
-                                  >
-                                    <div className="flex items-start gap-3">
-                                      <div className="size-10 rounded-xl bg-zinc-100 text-zinc-500 flex items-center justify-center shrink-0">
-                                        <Building2 className="size-4" />
-                                      </div>
-                                      <div>
-                                        <p className="text-[10px] font-semibold uppercase tracking-widest text-zinc-400 mb-1">
-                                          Organization
-                                        </p>
-                                        {selectedDonor.organization && (
-                                          <p className="text-sm font-medium text-zinc-900">
-                                            {selectedDonor.organization}
-                                          </p>
-                                        )}
-                                        {selectedDonor.title && (
-                                          <p className="text-sm text-zinc-500">
-                                            {selectedDonor.title}
-                                          </p>
-                                        )}
-                                      </div>
-                                    </div>
-                                  </motion.div>
-                                )}
-
-                                <div className="grid grid-cols-2 gap-3">
-                                  {selectedDonor.spouse && (
-                                    <motion.div
-                                      variants={fadeInUp}
-                                      whileHover={{ y: -2 }}
-                                      className="p-4 bg-zinc-50 rounded-2xl border border-zinc-100"
-                                    >
-                                      <div className="flex items-center gap-3">
-                                        <div className="size-10 rounded-xl bg-rose-50 text-rose-500 flex items-center justify-center shrink-0">
-                                          <Heart className="size-4" />
-                                        </div>
-                                        <div>
-                                          <p className="text-[10px] font-semibold uppercase tracking-widest text-zinc-400">
-                                            Spouse
-                                          </p>
-                                          <p className="text-sm font-medium text-zinc-900">
-                                            {selectedDonor.spouse}
-                                          </p>
-                                        </div>
-                                      </div>
-                                    </motion.div>
-                                  )}
-                                  {selectedDonor.birthday && (
-                                    <motion.div
-                                      variants={fadeInUp}
-                                      whileHover={{ y: -2 }}
-                                      className="p-4 bg-zinc-50 rounded-2xl border border-zinc-100"
-                                    >
-                                      <div className="flex items-center gap-3">
-                                        <div className="size-10 rounded-xl bg-amber-50 text-amber-500 flex items-center justify-center shrink-0">
-                                          <Star className="size-4" />
-                                        </div>
-                                        <div>
-                                          <p className="text-[10px] font-semibold uppercase tracking-widest text-zinc-400">
-                                            Birthday
-                                          </p>
-                                          <p className="text-sm font-medium text-zinc-900">
-                                            {format(
-                                              parseDisplayDate(
-                                                selectedDonor.birthday,
-                                              ),
-                                              "MMMM d",
-                                            )}
-                                          </p>
-                                        </div>
-                                      </div>
-                                    </motion.div>
-                                  )}
-                                  {selectedDonor.anniversary && (
-                                    <motion.div
-                                      variants={fadeInUp}
-                                      whileHover={{ y: -2 }}
-                                      className="p-4 bg-zinc-50 rounded-2xl border border-zinc-100"
-                                    >
-                                      <div className="flex items-center gap-3">
-                                        <div className="size-10 rounded-xl bg-purple-50 text-purple-500 flex items-center justify-center shrink-0">
-                                          <Calendar className="size-4" />
-                                        </div>
-                                        <div>
-                                          <p className="text-[10px] font-semibold uppercase tracking-widest text-zinc-400">
-                                            Anniversary
-                                          </p>
-                                          <p className="text-sm font-medium text-zinc-900">
-                                            {format(
-                                              parseDisplayDate(
-                                                selectedDonor.anniversary,
-                                              ),
-                                              "MMMM d",
-                                            )}
-                                          </p>
-                                        </div>
-                                      </div>
-                                    </motion.div>
-                                  )}
-                                </div>
-
-                                {selectedDonor.notes && (
-                                  <motion.div
-                                    variants={fadeInUp}
-                                    whileHover={{ y: -2 }}
-                                    className="p-4 bg-amber-50/50 rounded-2xl border border-amber-100"
-                                  >
-                                    <p className="text-[10px] font-semibold uppercase tracking-widest text-amber-600 mb-2">
-                                      Internal Notes
-                                    </p>
-                                    <p className="text-sm text-zinc-700">
-                                      {selectedDonor.notes}
-                                    </p>
-                                  </motion.div>
-                                )}
-                              </motion.div>
-                            </div>
-                          </TabsContent>
-
-                          <TabsContent
-                            value="recurring"
-                            className="mt-0 space-y-6"
-                          >
-                            <motion.div
-                              {...fadeInUp}
-                              transition={smoothTransition}
-                              className="flex items-center justify-between mb-2"
-                            >
-                              <div>
-                                <h3 className="text-sm font-semibold text-zinc-900">
-                                  Recurring Donations
-                                </h3>
-                                <p className="text-xs text-zinc-500 mt-0.5">
-                                  Scheduled giving commitments for this partner
-                                </p>
-                              </div>
-                            </motion.div>
-
-                            {selectedDonor.recurring_donations.length === 0 ? (
-                              <motion.div
-                                {...fadeInUp}
-                                className="flex flex-col items-center justify-center py-16 text-center bg-zinc-50 rounded-2xl border border-zinc-100"
-                              >
-                                <motion.div
-                                  initial={{ scale: 0.8 }}
-                                  animate={{ scale: 1 }}
-                                  transition={springTransition}
-                                  className="size-16 bg-white rounded-2xl flex items-center justify-center mb-4 shadow-sm"
-                                >
-                                  <Repeat className="size-7 text-zinc-300" />
-                                </motion.div>
-                                <p className="text-sm font-semibold text-zinc-900">
-                                  No recurring donations
-                                </p>
-                                <p className="text-xs text-zinc-400 mt-1 max-w-[280px]">
-                                  When this partner sets up a recurring gift, it
-                                  will appear here with all the details.
-                                </p>
-                              </motion.div>
-                            ) : (
-                              <motion.div
-                                variants={staggerContainer}
-                                initial="initial"
-                                animate="animate"
-                                className="space-y-4"
-                              >
-                                {selectedDonor.recurring_donations.map(
-                                  (recurring, i) => (
-                                    <motion.div
-                                      key={recurring.id}
-                                      variants={fadeInUp}
-                                      transition={{ delay: i * 0.1 }}
-                                      whileHover={{ y: -2 }}
-                                      className={cn(
-                                        "p-5 rounded-2xl border transition-[color,background-color,border-color,box-shadow,transform,opacity]",
-                                        recurring.status === "active"
-                                          ? "bg-linear-to-br from-emerald-50/80 to-emerald-50/30 border-emerald-200"
-                                          : "bg-zinc-50 border-zinc-200",
-                                      )}
-                                    >
-                                      <div className="flex flex-col lg:flex-row lg:items-start justify-between gap-4 mb-5">
-                                        <div className="flex items-start gap-4">
-                                          <motion.div
-                                            whileHover={{
-                                              scale: 1.1,
-                                              rotate: 5,
-                                            }}
-                                            className={cn(
-                                              "size-12 rounded-xl flex items-center justify-center shrink-0",
-                                              recurring.status === "active"
-                                                ? "bg-emerald-100"
-                                                : "bg-zinc-100",
-                                            )}
-                                          >
-                                            {getPaymentMethodIcon(
-                                              recurring.payment_method,
-                                            )}
-                                          </motion.div>
-                                          <div>
-                                            <div className="flex items-center gap-3 mb-1">
-                                              <h4 className="text-xl font-semibold text-zinc-900">
-                                                {formatCurrency(
-                                                  Number(recurring.amount),
-                                                )}
-                                              </h4>
-                                              <span className="text-sm font-medium text-zinc-500">
-                                                /{" "}
-                                                {recurring.frequency.toLowerCase()}
-                                              </span>
-                                              {getRecurringStatusBadge(
-                                                recurring.status as RecurringStatus,
-                                              )}
-                                            </div>
-                                            <div className="flex flex-wrap items-center gap-x-4 gap-y-1 text-xs text-zinc-500">
-                                              <span className="flex items-center gap-1">
-                                                <Calendar className="size-3.5" />
-                                                Started{" "}
-                                                {format(
-                                                  parseDisplayDate(
-                                                    recurring.start_date,
-                                                  ),
-                                                  "MMM d, yyyy",
-                                                )}
-                                              </span>
-                                              {recurring.end_date ? (
-                                                <span className="flex items-center gap-1 text-amber-600">
-                                                  <Clock className="size-3.5" />
-                                                  Ends{" "}
-                                                  {format(
-                                                    parseDisplayDate(
-                                                      recurring.end_date,
-                                                    ),
-                                                    "MMM d, yyyy",
-                                                  )}
-                                                </span>
-                                              ) : (
-                                                <span className="text-emerald-600">
-                                                  No end date
-                                                </span>
-                                              )}
-                                            </div>
-                                          </div>
-                                        </div>
-                                        {recurring.status === "active" &&
-                                          recurring.next_payment_date && (
-                                            <motion.div
-                                              initial={{
-                                                opacity: 0,
-                                                scale: 0.9,
-                                              }}
-                                              animate={{ opacity: 1, scale: 1 }}
-                                              className="bg-white p-3 rounded-xl border border-emerald-100 text-center lg:text-right"
-                                            >
-                                              <p className="text-[10px] font-semibold uppercase tracking-widest text-zinc-400">
-                                                Next Payment
-                                              </p>
-                                              <p className="text-lg font-semibold text-zinc-900">
-                                                {format(
-                                                  parseDisplayDate(
-                                                    recurring.next_payment_date,
-                                                  ),
-                                                  "MMM d",
-                                                )}
-                                              </p>
-                                              <p className="text-xs text-zinc-500">
-                                                {formatDistanceToNow(
-                                                  parseDisplayDate(
-                                                    recurring.next_payment_date,
-                                                  ),
-                                                  { addSuffix: true },
-                                                )}
-                                              </p>
-                                            </motion.div>
-                                          )}
-                                      </div>
-
-                                      <div className="grid grid-cols-2 sm:grid-cols-5 gap-4 p-4 bg-white/60 rounded-xl border border-zinc-100">
-                                        {[
-                                          {
-                                            label: "Payment Method",
-                                            value:
-                                              recurring.payment_method ||
-                                              "Online",
-                                            icon: true,
-                                          },
-                                          {
-                                            label: "Total Paid",
-                                            value: formatCurrency(
-                                              Number(recurring.total_paid),
-                                            ),
-                                            color: "text-emerald-600",
-                                          },
-                                          {
-                                            label: "Expected",
-                                            value: formatCurrency(
-                                              Number(recurring.total_expected),
-                                            ),
-                                          },
-                                          {
-                                            label: "Completed",
-                                            value: `${recurring.payments_completed} payments`,
-                                          },
-                                          {
-                                            label: "Remaining",
-                                            value:
-                                              recurring.payments_remaining > 0
-                                                ? `${recurring.payments_remaining} payments`
-                                                : "Ongoing",
-                                          },
-                                        ].map((item) => (
-                                          <div key={item.label}>
-                                            <p className="text-[10px] font-semibold uppercase tracking-widest text-zinc-400 mb-1">
-                                              {item.label}
-                                            </p>
-                                            <div className="flex items-center gap-1.5">
-                                              {item.icon &&
-                                                getPaymentMethodIcon(
-                                                  recurring.payment_method,
-                                                )}
-                                              <p
-                                                className={cn(
-                                                  "text-sm font-semibold",
-                                                  item.color || "text-zinc-900",
-                                                )}
-                                              >
-                                                {item.value}
-                                              </p>
-                                            </div>
-                                          </div>
-                                        ))}
-                                      </div>
-
-                                      <div className="mt-4">
-                                        <div className="flex items-center justify-between mb-1.5">
-                                          <span className="text-[10px] font-semibold uppercase tracking-widest text-zinc-400">
-                                            Progress
-                                          </span>
-                                          <span className="text-xs font-semibold text-zinc-600">
-                                            {Number(recurring.total_expected) >
-                                            0
-                                              ? `${Math.round((Number(recurring.total_paid) / Number(recurring.total_expected)) * 100)}%`
-                                              : "Ongoing"}
-                                          </span>
-                                        </div>
-                                        <div className="h-2 bg-zinc-200 rounded-full overflow-hidden">
-                                          <motion.div
-                                            initial={{ width: 0 }}
-                                            animate={{
-                                              width:
-                                                Number(
-                                                  recurring.total_expected,
-                                                ) > 0
-                                                  ? `${Math.min((Number(recurring.total_paid) / Number(recurring.total_expected)) * 100, 100)}%`
-                                                  : "100%",
-                                            }}
-                                            transition={{
-                                              duration: 0.8,
-                                              ease: "easeOut",
-                                            }}
-                                            className={cn(
-                                              "h-full rounded-full",
-                                              recurring.status === "active"
-                                                ? "bg-emerald-500"
-                                                : recurring.status ===
-                                                    "completed"
-                                                  ? "bg-blue-500"
-                                                  : "bg-zinc-400",
-                                            )}
-                                          />
-                                        </div>
-                                      </div>
-                                    </motion.div>
-                                  ),
-                                )}
-                              </motion.div>
-                            )}
-                          </TabsContent>
-
-                          <TabsContent value="giving" className="mt-0">
-                            <DataTableResponsive
-                              columns={givingHistoryColumns}
-                              data={givingHistoryRows}
-                              config={{
-                                enableRowSelection: false,
-                                enableColumnVisibility: false,
-                                enablePagination: true,
-                                enableFilters: false,
-                                enableSorting: true,
-                              }}
-                              emptyState={
-                                <div className="flex flex-col items-center justify-center py-12 text-center">
-                                  <p className="text-sm font-semibold text-zinc-900">
-                                    No giving history available
-                                  </p>
-                                  <p className="text-xs text-zinc-400 mt-1">
-                                    Gift activity will appear here once
-                                    donations are recorded.
-                                  </p>
-                                </div>
-                              }
-                            />
-                          </TabsContent>
-                        </AnimatePresence>
-                      </div>
-                    </ScrollArea>
-                  </Tabs>
-                </Card>
-              </motion.div>
-            ) : (
-              <motion.div
-                key="empty"
-                {...scaleIn}
-                transition={smoothTransition}
-              >
-                <Card className="border-zinc-200 border-dashed bg-zinc-50/30 rounded-[2.5rem] h-full min-h-[600px] flex items-center justify-center">
-                  <CardContent className="p-16 text-center">
-                    <motion.div
-                      initial={{ scale: 0.8, opacity: 0 }}
-                      animate={{ scale: 1, opacity: 1 }}
-                      transition={springTransition}
-                      className="size-20 rounded-3xl bg-white shadow-sm border border-zinc-100 flex items-center justify-center mx-auto mb-8"
-                    >
-                      <User className="size-10 text-zinc-200" />
-                    </motion.div>
-                    <motion.h3
-                      initial={{ opacity: 0, y: 10 }}
-                      animate={{ opacity: 1, y: 0 }}
-                      transition={{ delay: 0.1 }}
-                      className="font-semibold text-2xl text-zinc-900 tracking-tight"
-                    >
-                      Select a Partner
-                    </motion.h3>
-                    <motion.p
-                      initial={{ opacity: 0, y: 10 }}
-                      animate={{ opacity: 1, y: 0 }}
-                      transition={{ delay: 0.15 }}
-                      className="mt-2 text-sm font-medium text-zinc-400 max-w-[280px] mx-auto"
-                    >
-                      Choose a donor from the list to view their profile,
-                      recurring donations, and giving history.
-                    </motion.p>
-                    {profile?.id && (
-                      <motion.div
-                        initial={{ opacity: 0, y: 10 }}
-                        animate={{ opacity: 1, y: 0 }}
-                        transition={{ delay: 0.2 }}
-                        whileHover={{ scale: 1.02 }}
-                        whileTap={{ scale: 0.98 }}
-                      >
-                        <AddPartnerDialog
-                          missionaryId={profile.id}
-                          onSuccess={refreshDonors}
-                          trigger={
-                            <Button className="mt-10 h-11 px-8 rounded-2xl bg-zinc-900 text-[10px] font-semibold uppercase tracking-[0.2em] text-white hover:bg-zinc-800">
-                              <Plus className="size-4 mr-2" /> Add Partner
-                            </Button>
-                          }
-                        />
-                      </motion.div>
-                    )}
-                  </CardContent>
-                </Card>
-              </motion.div>
-            )}
-          </AnimatePresence>
-        </motion.div>
-      </div>
-
-      <Dialog
-        open={noteComposer.isOpen}
-        onOpenChange={(open) => {
-          if (open) {
-            noteComposer.open();
-            return;
-          }
-
-          noteComposer.close();
-        }}
-      >
-        <DialogContent className="sm:max-w-[500px] rounded-2xl">
-          <DialogHeader>
-            <DialogTitle className="text-lg font-semibold tracking-tight">
-              {noteComposer.activityType === "note"
-                ? "Add Note"
-                : noteComposer.activityType === "call"
-                  ? "Log Call"
-                  : noteComposer.activityType === "meeting"
-                    ? "Log Meeting"
-                    : "Log Email"}
-            </DialogTitle>
-            <DialogDescription className="text-sm text-zinc-500">
-              Add to {selectedDonor?.name}&apos;s timeline.
-            </DialogDescription>
-          </DialogHeader>
-          <div className="py-4">
-            <Textarea
-              value={noteComposer.noteInput}
-              onChange={(e) => noteComposer.setNoteInput(e.target.value)}
-              placeholder={
-                noteComposer.activityType === "call"
-                  ? "What did you discuss?"
-                  : noteComposer.activityType === "meeting"
-                    ? "Meeting notes..."
-                    : "Type your note here..."
-              }
-              className="min-h-[150px] resize-none rounded-xl border-zinc-200"
-            />
-          </div>
-          <DialogFooter className="gap-2 sm:gap-0">
-            <Button
-              variant="outline"
-              onClick={noteComposer.close}
-              className="h-10 px-6 rounded-xl border-zinc-200"
-            >
-              Cancel
-            </Button>
-            <Button
-              onClick={noteComposer.save}
-              disabled={!noteComposer.noteInput.trim() || noteComposer.isSaving}
-              className="h-10 px-6 rounded-xl"
-            >
-              {noteComposer.isSaving ? (
-                <Loader2 className="size-4 animate-spin" />
-              ) : (
-                "Save"
-              )}
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
-
-      <Dialog
-        open={tagEditor.isOpen}
-        onOpenChange={(open) => {
-          if (open) {
-            tagEditor.open();
-            return;
-          }
-
-          tagEditor.close();
-        }}
-      >
-        <DialogContent className="sm:max-w-[500px] rounded-2xl">
-          <DialogHeader>
-            <DialogTitle className="text-lg font-semibold tracking-tight">
-              Manage Tags
-            </DialogTitle>
-            <DialogDescription className="text-sm text-zinc-500">
-              Select tags for {selectedDonor?.name}. Tags help you organize and
-              filter your partners.
-            </DialogDescription>
-          </DialogHeader>
-          <div className="py-4">
-            <motion.div
-              variants={staggerContainer}
-              initial="initial"
-              animate="animate"
-              className="flex flex-wrap gap-2"
-            >
-              {AVAILABLE_TAGS.map((tag, i) => (
-                <motion.button
-                  key={tag.id}
-                  variants={fadeInUp}
-                  transition={{ delay: i * 0.02 }}
-                  whileHover={{ scale: 1.02 }}
-                  whileTap={{ scale: 0.98 }}
-                  onClick={() => tagEditor.toggleTag(tag.id)}
-                  className={cn(
-                    "px-3 py-1.5 rounded-full text-xs font-semibold border transition-[color,background-color,border-color,box-shadow,transform,opacity]",
-                    tagEditor.selectedTags.includes(tag.id)
-                      ? cn(tag.color, "ring-2 ring-offset-1 ring-zinc-400")
-                      : "bg-zinc-50 text-zinc-400 border-zinc-200 hover:bg-zinc-100",
-                  )}
-                >
-                  <AnimatePresence mode="wait">
-                    {tagEditor.selectedTags.includes(tag.id) && (
-                      <motion.span
-                        initial={{ width: 0, opacity: 0 }}
-                        animate={{ width: "auto", opacity: 1 }}
-                        exit={{ width: 0, opacity: 0 }}
-                        className="inline-flex overflow-hidden"
-                      >
-                        <Check className="size-3 mr-1" />
-                      </motion.span>
-                    )}
-                  </AnimatePresence>
-                  {tag.label}
-                </motion.button>
-              ))}
-            </motion.div>
-          </div>
-          <DialogFooter className="gap-2 sm:gap-0">
-            <Button
-              variant="outline"
-              onClick={tagEditor.close}
-              className="h-10 px-6 rounded-xl border-zinc-200"
-            >
-              Cancel
-            </Button>
-            <Button
-              onClick={tagEditor.save}
-              disabled={tagEditor.isSaving}
-              className="h-10 px-6 rounded-xl"
-            >
-              {tagEditor.isSaving ? (
-                <Loader2 className="size-4 animate-spin" />
-              ) : (
-                "Save Tags"
-              )}
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
-
-      <EditDonorDialog
-        donor={selectedDonor}
-        onOpenChange={(open) => {
-          if (open) {
-            editDialog.open();
-            return;
-          }
-
-          editDialog.close();
-        }}
-        onSuccess={refreshDonors}
-        open={editDialog.isOpen}
-      />
-    </motion.div>
+    <DonorsPageViewContext.Provider value={view}>
+      {children}
+    </DonorsPageViewContext.Provider>
   );
+}
+
+export function useDonorsPageViewFields() {
+  const view = React.use(DonorsPageViewContext);
+  if (!view) {
+    throw new Error(
+      "useDonorsPageViewFields must be used within DonorsPageViewProvider",
+    );
+  }
+  return view;
 }
