+++ docs/guides/development/react-doctor.md
@@ -64,5 +64,5 @@ The ignore list is intentionally human-readable here because `doctor.config.json
 ## 2026-05-21 Cleanup Decisions
 
 - Responsive shared tables: `packages/ui/components/shadcn/data-table/data-table-responsive-inner.tsx` now derives mobile state with `useMediaQuery` and keeps only the table-to-card coercion effect. Avoid reintroducing manual `resize` listeners in table components unless the shared hook cannot express the query.
-- Missionary donor mutations: `apps/missionary/app/donors/use-donors-page-view.tsx` and `apps/missionary/app/donors/edit-donor-dialog.tsx` now call `/api/missionary/donors/[donorId]` and `/api/missionary/donors/[donorId]/activities` through `apps/missionary/app/donors/donor-mutation-client.ts`. Server-side handlers live in `packages/api/src/missionary-portal/donor.ts` and must keep missionary role, tenant, profile, and donor relationship scoping.
+- Missionary donor mutations: `apps/missionary/app/donors/use-donors-page-view.tsx` and `apps/missionary/app/donors/edit-donor-dialog.tsx` call `/api/missionary/donors/[donorId]` and `/api/missionary/donors/[donorId]/activities` through `apps/missionary/app/donors/donor-mutation-client.ts`. Roster, detail, and dialog adapters consume the page view-model and must not call the mutation client or the missionary donor query directly. Server-side handlers live in `packages/api/src/missionary-portal/donor.ts` and must keep missionary role, tenant, profile, and donor relationship scoping.
 - Large component, bundle, accessibility, and design rule families remain per-slice work. Do not re-enable those globally without route-owner coverage and rendered validation.
