+++ tests/unit/apps/missionary/donors-page-view-model-contract.test.ts
@@ -22,4 +22,16 @@ describe("missionary donors page view-model contract", () => {
     expect(source).not.toContain("export type DonorsPageViewModel");
     expect(source).not.toContain('ReturnType<typeof useAuth>["profile"]');
   });
+
+  it("keeps table column factories out of the page view-model", () => {
+    const source = readRepoFile(
+      "apps/missionary/app/donors/use-donors-page-view.tsx",
+    );
+
+    expect(source).not.toContain("tables:");
+    expect(source).not.toContain("ColumnDef");
+    expect(source).not.toContain("DataTableResponsive");
+    expect(source).not.toContain("createDonorColumns");
+    expect(source).not.toContain("createGivingHistoryColumns");
+  });
 });
